REL-2412 Split bundle edu.gemini.dataman.app to support reusing the gsa query methods #time 6h
