+++ bundle/edu.gemini.dataman.app/src/main/scala/edu/gemini/dataman/app/GsaPostAction.scala
@@ -1,7 +1,9 @@
 package edu.gemini.dataman.app
 
 import edu.gemini.dataman.core._
-import edu.gemini.dataman.query.{GsaQaUpdateQuery, GsaResponse}
+import edu.gemini.gsa.core.{QaResponse, QaRequest}
+
+import edu.gemini.gsa.query.{GsaResponse, GsaAuth, GsaQaUpdateQuery, GsaHost}
 import edu.gemini.pot.spdb.IDBDatabaseService
 import edu.gemini.spModel.core.Site
 import edu.gemini.spModel.dataset.QaRequestStatus.ProcessingPost
