+++ bundle/edu.gemini.dataman.app/src/main/scala/edu/gemini/dataman/app/ObsLogActions.scala
@@ -2,6 +2,7 @@ package edu.gemini.dataman.app
 
 import edu.gemini.dataman.app.ModelActions._
 import edu.gemini.dataman.core._
+import edu.gemini.gsa.core.{QaRequest, GsaRecord}
 import edu.gemini.pot.sp.SPObservationID
 import edu.gemini.pot.spdb.IDBDatabaseService
 import edu.gemini.spModel.dataset.{DatasetQaState, Dataset, DatasetQaRecord, DatasetLabel, SummitState, DatasetExecRecord, DatasetGsaState}
