+++ bundle/edu.gemini.dataman.app/src/main/scala/edu/gemini/dataman/osgi/GsaCommands.scala
@@ -2,7 +2,8 @@ package edu.gemini.dataman.osgi
 
 import edu.gemini.dataman.app.{DmanActionExec, GsaPollActions}
 import edu.gemini.dataman.core._
-import edu.gemini.dataman.query.{GsaRecordQuery, GsaQaUpdateQuery, GsaResponse, TimeFormat}
+import edu.gemini.gsa.query.{GsaRecordQuery, GsaQaUpdateQuery, GsaResponse, TimeFormat}
+import edu.gemini.gsa.core.{QaResponse, QaRequest, GsaRecord}
 import edu.gemini.pot.spdb.IDBDatabaseService
 import edu.gemini.spModel.core.catchingNonFatal
 import edu.gemini.spModel.dataset.{DatasetQaState, DatasetLabel}
