+++ bundle/edu.gemini.gsa.query/src/main/scala/edu/gemini/gsa/core/GsaRecord.scala
@@ -1,4 +1,4 @@
-package edu.gemini.dataman.core
+package edu.gemini.gsa.core
 
 import edu.gemini.spModel.dataset.{DatasetGsaState, DatasetLabel}
