+++ bundle/edu.gemini.gsa.query/src/main/scala/edu/gemini/gsa/query/GsaHost.scala
@@ -1,4 +1,4 @@
-package edu.gemini.dataman.core
+package edu.gemini.gsa.query
 
 import scalaz._
