+++ bundle/edu.gemini.gsa.query/src/test/scala/edu/gemini/gsa/query/GsaQaUpdateQuerySpec.scala
@@ -1,6 +1,6 @@
-package edu.gemini.dataman.query
+package edu.gemini.gsa.query
 
-import edu.gemini.dataman.core.{QaResponse, QaRequest, Arbitraries}
+import edu.gemini.gsa.core.{QaResponse, QaRequest, Arbitraries}
 import edu.gemini.spModel.dataset.DatasetLabel
 import org.scalacheck._
 import org.scalacheck.Arbitrary._
